% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchDST.R
\name{fetch_DST}
\alias{fetch_DST}
\title{Get DST data}
\usage{
fetch_DST(agravo)
}
\arguments{
\item{agravo}{Character: 'AIDS','SIFI','HEPA','HEPB','HEPC','HEPD'}
}
\description{
Get DST data from clauciorank github: https://github.com/clauciorank/sinanBRFiles
the original ones could be find at:
http://indicadoreshepatites.aids.gov.br/;
http://indicadoressifilis.aids.gov.br/;
http://indicadores.aids.gov.br/;
}
\examples{
data <- fetch_DST('AIDS')
}
