% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSINAN.R
\name{fetch_SINAN}
\alias{fetch_SINAN}
\title{Get data from SINAN}
\usage{
fetch_SINAN(anomin, anomax, agravo, uf)
}
\arguments{
\item{anomin}{ano mínimo numérico: 2015,2020}

\item{anomax}{ano máximo numérico: 2015,2020}

\item{agravo}{Character: sigla do agravo Acidente por Animais Peçonhentos (ANIM), Botulismo (BOTU), Doença de Chagas (CHAG), Febre de Chikungunya (CHIK), Cólera (COLE), Coqueluche (COQU), Dengue (DENG), Difteria (DIFT), Esquistossomose (ESQU), Febre Amarela (FAMA), Febre Maculosa (FMAC), Febre Tifóide (FTIF),Hanseníase (HANS), Hantavirose (HANT),Intoxicação Exógena (IEXO), Influenza Pandêmica (INFL), Leishmaniose Visceral (LEIV),    Leptospirose (LEPT), Leishmaniose Tegumentar Americana (LTAN), Malária (MALA), Zika Vírus (ZIKA)}

\item{uf}{character sigla estado: 'SC', 'PR', 'ALL'}
}
\description{
Get data from SINAN
}
\examples{
data <- fetch_SINAN(2018, 2020, 'DENG', 'ALL')
data <- fetch_SINAN(2015, 2020, 'HANS', 'TO')
}
